add get_exercise route
