+++ src/models/base.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import DeclarativeBase, sessionmaker
 from sqlalchemy import create_engine
 
-engine = create_engine("sqlite:///test.db", echo=True)
+engine = create_engine("sqlite:///test.db")
 
 Session = sessionmaker(engine)
